watchmate/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-filter==24.2
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
